Some initializations
